tests/wpt/tests/html/semantics/embedded-content/the-img-element/naturalWidth-naturalHeight-unavailable.tentative.html

<!doctype html>
<head>
<meta charset="utf-8">
<title>HTMLImageElement naturalWidth/naturalHeight behavior for SVG that lacks at least one natural dimension</title>
<!-- Note: this test asserts a different expectation from what the HTML spec
     requires, as of mid-2025 when this testcase is being written. The spec
     behavior doesn't appear to be web-compatible for some of the cases here,
     and issue https://github.com/whatwg/html/issues/11287 is filed on
     addresing that.  In the meantime, this test is named with ".tentative" to
     indicate that it's not authoritative. After the spec change is accepted,
     we can remove the neighboring naturalWidth-naturalHeight.html test which
     asserts the prior spec text's expectations, since this test covers the
     same ground but with its expectations set according to the
     soon-to-be-updated spec text.  -->
<link rel="help" href="https://github.com/whatwg/html/issues/11287">
<link rel="help" href="https://html.spec.whatwg.org/multipage/embedded-content.html#dom-img-naturalwidth-dev">
<link rel="help" href="https://html.spec.whatwg.org/multipage/embedded-content.html#dom-img-width">
<link rel="help" href="https://html.spec.whatwg.org/multipage/images.html#density-corrected-intrinsic-width-and-height">
<link rel="help" href="https://drafts.csswg.org/css-images/#natural-dimensions">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#scroller {
  /* We wrap all the test content in a scroller so that it doesn't push
   * the textual test-results too far out of view.
   */
  border: 1px solid black;
  height: 400px;
  width: max-content;
  overflow: scroll;
}
#containingBlock {
  /* There are a few SVG images here that size so that their margin-box fills
   * their containing block width. We define a specific size here so that we
   * can then check for it (minus the margins) in the "data-width" attribute.
   */
  width: 740px;
}
img {
  /* This styling is just cosmetic, to help visualize the images. */
  border: 5px solid teal;
  margin: 5px;
  display: block;
}
</style>
<!-- We specify the img elements in a <template> and then clone them for
     testing, so that we can dynamically generate and test several variants
     of each img. -->
<template id="imgTemplates">
<!-- For each img element:
     * The "data-natural-{width,height}" attributes represent the expected
     values of the img element's "naturalWidth" and "naturalHeight" IDL
     attributes. This test implicitly expects the "width" and "height" IDL
     attributes to have those same expected values; but in cases where that's
     not correct, we provide the actual expected value in the
     "data-{width,height}" attributes (as distinguished from
     data-natural-{width,height}).
     * The "title" attribute is a description of the scenario being tested, and
     it must be unique to satisfy the test harness requirements. -->

<!-- FIRST PART OF TEST: No viewBox. Just a missing (or edge-casey, i.e.
     negative or percent-valued) values, for the width and height attrs on the
     root svg element in a SVG image. -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg'></svg>"
     title="SVG image, no natural dimensions"
     data-natural-width="300" data-natural-height="150">
<!-- Note: percent values can't be resolved when determining natural
     dimensions, so the exact percentage shouldn't matter. -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='400%' height='10%'></svg>"
     title="SVG image, percengage natural dimensions"
     data-natural-width="300" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='-400%' height='-10%'></svg>"
     title="SVG image, negative percengage natural dimensions"
     data-natural-width="300" data-natural-height="150">
<!-- If only one attribute is present, it should show up as a natural
     dimension, without influencing the other natural dimension. -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='60'></svg>"
     title="SVG image, with natural width"
     data-natural-width="60" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='60'></svg>"
     title="SVG image, with natural height"
     data-natural-width="300" data-natural-height="60">
<!-- If either attribute is 0 or a negative length, it should show up as a
     natural dimension: of 0. -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='0'></svg>"
     title="SVG image, with natural width of 0"
     data-natural-width="0" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='0'></svg>"
     title="SVG image, with natural height of 0"
     data-natural-width="300" data-natural-height="0">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='-5'></svg>"
     title="SVG image, with natural width being negative"
     data-natural-width="0" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='-5'></svg>"
     title="SVG image, with natural height being negative"
     data-natural-width="300" data-natural-height="0">

<!-- NEXT PART OF TEST: Same as above, but now with a viewBox that grants a
     3:1 aspect-ratio; whenever we know one natural dimension, that should
     combine with the aspect ratio to produce the other natural dimension.

     NOTE: for a few subtests here, the image ends up expanding to fill the
     containing block's width, i.e. rendering at a larger size than its natural
     size. In those cases, we include 'data-width' & 'data-height' attributes,
     so that this test's JS can validate that img.width and img.height return
     these expected larger values. (Otherwise, we expect img.width and
     img.height to return the same values as img.naturalWidth and
     img.naturalHeight). -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 600 200'></svg>"
     title="SVG image, no natural dimensions, and aspect ratio from viewBox"
     data-natural-width="300" data-natural-height="100"
     data-width="720" data-height="240">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='400%' height='10%' viewBox='0 0 600 200'></svg>"
     title="SVG image, percengage natural dimensions, and aspect ratio from viewBox"
     data-natural-width="300" data-natural-height="100"
     data-width="720" data-height="240">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='-400%' height='-10%' viewBox='0 0 600 200'></svg>"
     title="SVG image, negative percengage natural dimensions, and aspect ratio from viewBox"
     data-natural-width="300" data-natural-height="100"
     data-width="720" data-height="240">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='60' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural width, and aspect ratio from viewBox"
     data-natural-width="60" data-natural-height="20">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='60' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural height, and aspect ratio from viewBox"
     data-natural-width="180" data-natural-height="60">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='0' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural width of 0, and aspect ratio from viewBox"
     data-natural-width="0" data-natural-height="0">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='0' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural height of 0, and aspect ratio from viewBox"
     data-natural-width="0" data-natural-height="0">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='-5' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural width being negative, and aspect ratio from viewBox"
     data-natural-width="0" data-natural-height="0">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='-5' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural height being negative, and aspect ratio from viewBox"
     data-natural-width="0" data-natural-height="0">

<!-- THIRD PART OF TEST: Check a degenerate 0-sized viewBox for some of the
     cases; it should have no impact. -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 0 0'></svg>"
     title="SVG image, no natural dimensions, viewBox with 0 width/height"
     data-natural-width="300" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 10 0'></svg>"
     title="SVG image, no natural dimensions, viewBox with 0 width"
     data-natural-width="300" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 0 10'></svg>"
     title="SVG image, no natural dimensions, viewBox with 0 height"
     data-natural-width="300" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='60' viewBox='0 0 0 0'></svg>"
     title="SVG image, with natural width, viewBox with 0 width/height"
     data-natural-width="60" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='60' viewBox='0 0 10 0'></svg>"
     title="SVG image, with natural width, viewBox with 0 width"
     data-natural-width="60" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='60' viewBox='0 0 0 10'></svg>"
     title="SVG image, with natural width, viewBox with 0 height"
     data-natural-width="60" data-natural-height="150">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='60' viewBox='0 0 0 0'></svg>"
     title="SVG image, with natural height, viewBox with 0 width/height"
     data-natural-width="300" data-natural-height="60">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='60' viewBox='0 0 10 0'></svg>"
     title="SVG image, with natural height, viewBox with 0 width"
     data-natural-width="300" data-natural-height="60">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' height='60' viewBox='0 0 0 10'></svg>"
     title="SVG image, with natural height, viewBox with 0 height"
     data-natural-width="300" data-natural-height="60">

<!~- FINAL PART OF TEST: we have width/height/viewBox all specified on the
     svg element. The width and height attrs should determine the natural
     dimensions, with no impact from viewBox. -->
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='60' height='60' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural width and height, and aspect ratio from viewBox"
     data-natural-width="60" data-natural-height="60">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='0' height='0' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural width and height of 0, and aspect ratio from viewBox"
     data-natural-width="0" data-natural-height="0">
<img src="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' width='-5' height='-5' viewBox='0 0 600 200'></svg>"
     title="SVG image, with natural width and height being negative, and aspect ratio from viewBox"
     data-natural-width="0" data-natural-height="0">
</template>
</head>
<body>
<div id="scroller">
  <div id="containingBlock">
  </div>
</div>
<!-- We generate and append all of the tested <img> elements while we're inside
     the <body>, so that all of the <img> elements' "load" events will block
     the window onload event: -->
<script>
setup({explicit_done:true});
// Clone and append a copy of the contents of imgTemplates, for testing:
let clone = imgTemplates.content.cloneNode("true");
containingBlock.appendChild(clone);

// After all the img elements have loaded (indicated by the window load event),
// we run the various tests:
onload = function() {
  Array.from(document.images).forEach(img => {
    test(function() {
      // We expect naturalWidth to match the provided data-natural-width
      // (and similar for 'height').
      const expectedNaturalWidth = parseFloat(img.dataset.naturalWidth);
      const expectedNaturalHeight = parseFloat(img.dataset.naturalHeight);
      assert_equals(img.naturalWidth, expectedNaturalWidth, 'naturalWidth');
      assert_equals(img.naturalHeight, expectedNaturalHeight, 'naturalHeight');

      // If 'data-width' is provided, then we expect img.width to match it.
      // Otherwise we expect img.width to match the 'data-natural-width'.
      // (And similar for 'height'.)
      const expectedWidth = 'width' in img.dataset ?
            parseFloat(img.dataset.width) : expectedNaturalWidth;
      const expectedHeight = 'height' in img.dataset ?
            parseFloat(img.dataset.height) : expectedNaturalHeight;
      assert_equals(img.width, expectedWidth, 'width');
      assert_equals(img.height, expectedHeight, 'height');

    }, `${img.title}`);
  });
  done();
};
</script>
</body>
